remove queries as `IntegrationTestSuite` to use as a static method
