+++ tests/e2e/e2e_distribution_test.go
@@ -22,7 +22,7 @@ func (s *IntegrationTestSuite) testDistribution(
 		beforeBalance = sdk.NewCoin(uatomDenom, math.NewInt(0))
 	}
 
-	s.execSetWithrawAddress(s.chainA, 0, fees.String(), delegatorAddress, newWithdrawalAddress, homePath)
+	s.execSetWithdrawAddress(s.chainA, 0, fees.String(), delegatorAddress, newWithdrawalAddress, homePath)
 
 	// Verify
 	s.Require().Eventually(
